refactor: refactor reasoning_parser lib, rename basic reasoning parser
